update pkg from /golang to /go. and complete the documentation
