+++ go/explode_test.go
@@ -1,10 +1,11 @@
 package pehape_test
 
 import (
+	"fmt"
 	"strings"
 	"testing"
 
-	PHP "github.com/teknologi-umum/pehape/golang"
+	PHP "github.com/teknologi-umum/pehape/go"
 )
 
 //function to test explode function
@@ -17,6 +18,7 @@ func TestExplode(t *testing.T) {
 
 	t.Run("Can Explode Space Separated String By Space", func(t *testing.T) {
 		str := "Hello pehape world"
+		fmt.Println(PHP.Explode(" ", str, -2))
 
 		//nil
 		if res, _ := PHP.Explode(" ", str); res[0] != []string{"Hello", "pehape", "world"}[0] {
