+++ go/go.mod
@@ -1,0 +1,3 @@
+module github.com/teknologi-umum/pehape/go
+
+go 1.18
