--- golang/go.mod
@@ -1,3 +1,0 @@
-module github.com/teknologi-umum/pehape/golang
-
-go 1.18
