--- golang/readme.md
@@ -1,25 +1,0 @@
-# Install Package
-
-Get package using
-`go get github.com/teknologi-umum/pehape/go`
-
-## List `Function` Already Which You Can Use
-
-### How to use?
- * `Implode`
-   `note` : `Don't forget to import pakage` => `import github.com/teknologi-umum/pehape/go`
-  - sample string   
-```go
-var aray = []string{"Hello", "world"}
-fmt.Println(Implode(array, " "))
-
-//result : Hello world
-```
-  - sample number
-```go
-var array = []int{1, 2, 3, 4, 5}
-fmt.Println(Implode(array, " "))
-
-//result : 1 2 3 4 5
-```
-
